% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particpants_responses.R
\docType{data}
\name{participants_responses}
\alias{participants_responses}
\title{Participants Responses}
\format{A data frame with 14001 rows and 1539 variables.}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/BPCDH5}
}
\usage{
participants_responses
}
\description{
I believe this is where we will find the variables we are interested in.
}
\keyword{datasets}
