% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_text.R
\docType{data}
\name{questions_text}
\alias{questions_text}
\title{Questions text}
\format{A dataframe with}
\usage{
questions_text
}
\description{
Had to do this the ol' fashioned way and munge it in a spreadsheet.
Provides the text of the question with the year and question no provided.
This is for political knowledge only.
}
\keyword{datasets}
