% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts.R
\docType{data}
\name{forecasts}
\alias{forecasts}
\title{Forecasts data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 448418 rows and 23 columns.}
\usage{
forecasts
}
\description{
To do!
}
\details{
See readme for deets and the codefile in raw-data/scripts/
}
\keyword{datasets}
