% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts.R
\docType{data}
\name{forecasts}
\alias{forecasts}
\title{Forecasts data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 935419 rows and 22 columns.}
\usage{
forecasts
}
\description{
See readme for deets and the codefile in raw-data/scripts/
}
\keyword{datasets}
